Part Six and Seven Complement
Fixed several problems, added a function in categorycontainer since the original format wasn't allowing the modification of a Category retrieved by a GetCategory.
Modified another problem in the function OpenRemoveCategory since there was a wrong index. I also changed a behaviour in the Game displayer, it was returning the games from the uncategorized category where it should've been returning a specific category.