Main commit
I think the main thing I did something different to Javid was on the shooting, I summed the player's momentum to the bullet's speed and I did this so that it would mainly apply when the ship is heading the same direction as it's momentum, so it won't actually affect when the ship is opposite to its momentum. All the of the above I did based only on the dot product.